+++ src/XBotCore.cpp
@@ -32,28 +32,75 @@
 //#include <Ethernet.h>
 
 XBot::XBotCore::XBotCore(const char* config_yaml) : 
-//     XBotEcat(config_yaml), 
     _path_to_config(config_yaml)
 {
+    //TODO fix thread stuff
+    // set thread name
+    //const YAML::Node& board_ctrl = root_cfg["x_bot_ecat"]; // TBD check that the Node is defined
+    //set_thread_name(board_ctrl["name"].as<std::string>()); // TBD check that name is defined
+    
+    set_thread_name("XBOT");
+    // set thread period - not periodic
+    task_period_t t;
+    memset(&t, 0, sizeof(t));
+    t.period = {0,1};
+    set_thread_period(t);
+    
+    // set thread priority
+    set_thread_priority();
+    
+    
+    //TODO use FactoryPattern
+    //use shared library
+    //robotInterface = new XBot::Ethernet(_path_to_config.c_str()); 
+    robotControlInterface = new XBot::XBotEcat(_path_to_config.c_str());
+}
 
-  //TODO use FactoryPattern
-  //use shared library
-  //robotInterface = new XBot::Ethernet(_path_to_config.c_str()); 
-  robotInterface = new XBot::XBotEcat(_path_to_config.c_str());
+void XBot::XBotCore::set_thread_name(std::string thread_name)
+{
+    // save the thread name
+    this->thread_name = thread_name;
+    // set thread name
+    name = this->thread_name.c_str();
+}
+
+std::string XBot::XBotCore::get_thread_name(void)
+{
+    return thread_name;
+}
+
+void XBot::XBotCore::set_thread_period(task_period_t t)
+{
+    period.task_time = t.task_time;
+    period.period = t.period;
+}
+
+void XBot::XBotCore::set_thread_priority()
+{
+
+    // set scheduler policy
+#ifdef __XENO__
+    schedpolicy = SCHED_FIFO;
+#else
+    schedpolicy = SCHED_OTHER;
+#endif
+    
+    // set scheduler priority and stacksize
+    priority = sched_get_priority_max(schedpolicy);
+    stacksize = 0; // not set stak size !!!! YOU COULD BECAME CRAZY !!!!!!!!!!!!
 }
 
 void XBot::XBotCore::th_init( void * ){
   
-  std::cout<<"INIT Thread"<<std::endl;
-  robotInterface->init();
+  robotControlInterface->init();
   control_init();
 }
 
 void XBot::XBotCore::th_loop( void * ){
   
-  robotInterface->recv_from_slave();
+  robotControlInterface->recv_from_slave();
   control_loop();
-  robotInterface->send_to_slave();
+  robotControlInterface->send_to_slave();
   
 }
 
@@ -62,10 +109,10 @@ void XBot::XBotCore::control_init(void)
     
     // create robot from config file and any map
     XBot::AnyMapPtr anymap = std::make_shared<XBot::AnyMap>();
-    std::shared_ptr<XBot::IXBotJoint> xbot_joint(robotInterface);
-    std::shared_ptr<XBot::IXBotFT> xbot_ft(robotInterface);
-    std::shared_ptr<XBot::IXBotIMU> xbot_imu(robotInterface);
-    std::shared_ptr<XBot::IXBotHand> xbot_hand(robotInterface);
+    std::shared_ptr<XBot::IXBotJoint> xbot_joint(robotControlInterface);
+    std::shared_ptr<XBot::IXBotFT> xbot_ft(robotControlInterface);
+    std::shared_ptr<XBot::IXBotIMU> xbot_imu(robotControlInterface);
+    std::shared_ptr<XBot::IXBotHand> xbot_hand(robotControlInterface);
     
     (*anymap)["XBotJoint"] = boost::any(xbot_joint);
     (*anymap)["XBotFT"] = boost::any(xbot_ft);
@@ -100,7 +147,6 @@ double XBot::XBotCore::get_time()
 
 int XBot::XBotCore::control_loop(void) 
 {    
-  std::cout<<"loop Thread"<<std::endl;
 //     std::cout << "laurenzi" << std::endl;
     _iter++;
     _pluginHandler->run();
